% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggparallel.R
\name{ggparallel}
\alias{ggparallel}
\title{A function for generating parallel plots with separate axis for each variable}
\usage{
ggparallel(
  p_data,
  tick_number = 5,
  tick_length = 0.1,
  label_nudge_x = -0.3,
  label_nudge_y = 0,
  label_text_size = 3.5,
  group_name_text_size = 11,
  digit_rounding = 2
)
}
\arguments{
\item{p_data}{A data frame containing plotting data.
with the first column containing the group/observation names,
and the other columns being descriptory variables for each group/observation.}

\item{tick_number}{Number of ticks/breaks on the vertical axis. Defaults to 5.}

\item{tick_length}{Length of ticks on the vertical axis. Defaults to 0.1.}

\item{label_nudge_x}{How much to nudge the vertical axis labels in the horizontal direction. Defaults to -0.3.}

\item{label_nudge_y}{How much to nudge the vertical axis labels in the vertical direction. Defaults to 0.}

\item{label_text_size}{The font size of vertical axis labels. Defaults to 3.5.}

\item{group_name_text_size}{The font size of horizontal axis labels. Defaults to 11.}

\item{digit_rounding}{To how many digits are the numbers rounded. Defaults to 2.}
}
\value{
A ggplot chart.
}
\description{
A function for generating parallel plots with separate axis for each variable
}
\examples{
mtcars_gr <- mtcars \%>\%
 tibble::rownames_to_column(var = "group") \%>\%
 tibble::as_tibble() \%>\%
 tail(3) \%>\%
 dplyr::select(1:7)

ggparallel(mtcars_gr)

iris_summary <- iris \%>\%
 dplyr::group_by(Species) \%>\%
 dplyr::summarise(across(everything(), mean))

ggparallel(iris_summary)
}
