% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspider.R
\name{ggspider}
\alias{ggspider}
\title{A function for generating radar plots with a shared scaled axis,
or multiple axis with real values displayed for each variable (spider chart).}
\usage{
ggspider(
  p_data,
  polygon = TRUE,
  scaled = FALSE,
  draw_axis = TRUE,
  background_color = "gray99",
  area_fill = TRUE,
  fill_opacity = 0.05,
  central_distance = 0.2,
  axis_name_offset = 0.2,
  digit_rounding = 2
)
}
\arguments{
\item{p_data}{A data frame containing plotting data.
with the first column containing the group/observation names,
and the other columns being descriptory variables for each group/observation.}

\item{polygon}{Whether to use circles or polygons as the background. Defaults to TRUE.}

\item{scaled}{Whether to display a single axis with scaled data or
individual axis (a radar or spider chart). Defaults to FALSE.}

\item{draw_axis}{Whether to draw variable axis or not. Defaults to TRUE.}

\item{background_color}{The background color of the chart. Defaults to "gray 99".}

\item{area_fill}{Whether or not to fill the shapes made by connecting data points. Defaults to TRUE.}

\item{fill_opacity}{How opaque are the shape fills. Defaults to 0.05.}

\item{central_distance}{The offset from the center of the radar chart. Defaults to 0.2.}

\item{axis_name_offset}{The axis name offset from the axis. Defaults to 0.2.}

\item{digit_rounding}{To how many digits are the numbers rounded. Defaults to 2.}
}
\value{
A ggplot chart.
}
\description{
A function for generating radar plots with a shared scaled axis,
or multiple axis with real values displayed for each variable (spider chart).
}
\examples{
mtcars_gr <- mtcars \%>\%
 tibble::rownames_to_column(var = "group") \%>\%
 tibble::as_tibble() \%>\%
 tail(3) \%>\%
dplyr::select(1:7)

ggspider(mtcars_gr, polygon = TRUE)

ggspider(mtcars_gr, polygon = FALSE)

ggspider(mtcars_gr, polygon = TRUE, scaled = TRUE)

ggspider(mtcars_gr, polygon = FALSE, scaled = TRUE)

iris_summary <- iris \%>\%
 dplyr::group_by(Species) \%>\%
 dplyr::summarise(across(everything(), mean))

ggspider(iris_summary)
}
