% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggspider.R
\name{ggspider}
\alias{ggspider}
\title{A function for generating radar/spider charts with a shared scaled axis,
or multiple axis with real values displayed for each variable (spider chart).}
\usage{
ggspider(
  p_data,
  ci_data = NULL,
  polygon = TRUE,
  scaled = FALSE,
  real = FALSE,
  draw_axis = TRUE,
  n_labels = 5,
  subset = NULL,
  reorder_axis = NULL,
  background_color = "gray99",
  area_fill = TRUE,
  fill_opacity = 0.05,
  central_distance = 0.2,
  axis_name_offset = 0.2,
  digit_rounding = 2,
  axis_label_font_size = NULL,
  axis_label_font_face = NULL,
  axis_name_font_size = NULL,
  axis_name_font_face = NULL
)
}
\arguments{
\item{p_data}{A data frame containing plotting data.
with the first column containing the group/observation names,
and the other columns being descriptory variables for each group/observation.}

\item{ci_data}{A data frame confidence interval data containing four columns.
The first column specifies the group, the second the feature, and the third and fourth left and
right confidence interval borders, respectively. See the vignette for an example. Defaults to NULL.}

\item{polygon}{Whether to use circles or polygons as the background. Defaults to TRUE.}

\item{scaled}{Whether to display a single axis with scaled data or
individual axis (a radar or spider chart). Defaults to FALSE.}

\item{real}{Whether to use the single scale for the entire input data and define the center circle
as zero. Defaults to FALSE.
If real is set to TRUE, central_distance is set to 0 and n_labels is set to 4.}

\item{draw_axis}{Whether to draw variable axis or not. Defaults to TRUE.}

\item{n_labels}{How many labels to use per axis. Defaults to 5.}

\item{subset}{Names of row identifiers to plot as a subset of data,
but with axis range specified using the full dataset. Useful when trying to compare subsets of data.}

\item{reorder_axis}{Names of features (columns) for which to reverse axis order from default min-max to max-min. Defaults to NULL.}

\item{background_color}{The background color of the chart. Defaults to "gray 99".}

\item{area_fill}{Whether or not to fill the shapes made by connecting data points. Defaults to TRUE.}

\item{fill_opacity}{How opaque are the shape fills. Defaults to 0.05.}

\item{central_distance}{The offset from the center of the radar chart. Defaults to 0.2.}

\item{axis_name_offset}{The axis name offset from the axis. Defaults to 0.2.}

\item{digit_rounding}{To how many digits are the numbers rounded. Defaults to 2.}

\item{axis_label_font_size}{Font size for the axis labels.}

\item{axis_label_font_face}{Font face for the axis labels.}

\item{axis_name_font_size}{Font size for the axis names.}

\item{axis_name_font_face}{Font face for the axis names.}
}
\value{
A ggplot based spider/radar chart.
}
\description{
A function for generating radar/spider charts with a shared scaled axis,
or multiple axis with real values displayed for each variable (spider chart).
}
\examples{
mtcars_gr <- mtcars \%>\%
 tibble::rownames_to_column(var = "group") \%>\%
 tibble::as_tibble() \%>\%
 tail(3) \%>\%
 dplyr::select(1:7)

ggspider(mtcars_gr, polygon = TRUE)

ggspider(mtcars_gr, polygon = FALSE)

ggspider(mtcars_gr, polygon = TRUE, scaled = TRUE)

ggspider(mtcars_gr, polygon = FALSE, scaled = TRUE)

ggspider(mtcars_gr, reorder_axis = c("mpg", "hp"))

iris_summary <- iris \%>\%
 dplyr::group_by(Species) \%>\%
 dplyr::summarise(across(everything(), mean))

ggspider(iris_summary)
ggspider(iris_summary, reorder_axis = c("Sepal.Length"))

iris_summary_ci <- iris \%>\%
 dplyr::group_by(Species) \%>\%
 dplyr::summarise(across(everything(), ~ 1.97*sd(.)/sqrt(n())))

iris_ci <- iris_summary \%>\%
 tidyr::pivot_longer(-1, names_to = "parameter", values_to = "mean") \%>\%
 dplyr::left_join(iris_summary_ci \%>\%
 tidyr::pivot_longer(-1, names_to = "parameter", values_to = "ci")) \%>\%
 dplyr::mutate(min = mean - ci, max = mean + ci) \%>\%
 dplyr::select(-mean, -ci)

ggspider(iris_summary, ci_data = iris_ci)
ggspider(iris_summary, ci_data = iris_ci, reorder_axis = c("Sepal.Length"))
}
